+++ src/sll/include/sll/_generated_help_text.h
@@ -1,5 +1,5 @@
-#ifndef __HELP_TEXT_H__
-#define __HELP_TEXT_H__ 1
+#ifndef __SLL__GENERATED_HELP_TEXT_H__
+#define __SLL__GENERATED_HELP_TEXT_H__ 1
 #include <sll/types.h>